requirements.txt
scrapy
python-docx
Pillow